pyproject.toml
[project]
name = "lebab"
version = "0.4.0"
description = "Document translator - Word and Powerpoint - from one language to another, respecting the document structure"
authors = ["Iwan van der Kleijn <iwanvanderkleijn@gmail.com>"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.19",
    "langchain-openai>=0.3.7",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-pptx>=1.0.2",
]

